+++ SplitonsCoffeeScript/controllers/ListProjectsController.ts
@@ -1,4 +1,5 @@
 ///<reference path="../linq/linq.d.ts"/>
+///<reference path="../external/bootbox.d.ts"/>
 ///<reference path="../angular.d.ts"/>
 ///<reference path="../dataObjects/Project.ts"/>
 ///<reference path="../dataObjects/Transaction.ts"/>
@@ -16,12 +17,22 @@ angular.module('splitonsApp').controller(
                     $location.path('/project/' + $scope.newProjectName + "/1").replace();
                 }
             };
+
             $scope.deleteProject = function (projectId) {
-                for (var index in $scope.projects) {
-                    if ($scope.projects[index].id == projectId) {
-                        $scope.projects.splice(index, 1);
-                        projectsFactory.deleteProject(projectId);
+                bootbox.confirm({
+                    size: 'small',
+                    message: "Your message here…",
+                    callback: function(result){
+                        if(result){
+                            for (var index in $scope.projects) {
+                                if ($scope.projects[index].id == projectId) {
+                                    $scope.projects.splice(index, 1);
+                                    projectsFactory.deleteProject(projectId);
+                                    $route.reload();
+                                }
+                            }
+                        }
                     }
-                }
+                });
             }
         }]);
